requirements.txt
# pip install -r requirements.txt
# python3.8 environment

# torch>=1.8.0
# torchvision>=0.9.0
torch==1.13.1
torchvision==0.14.1
numpy>=1.24.0
opencv-python>=4.1.2
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0
addict>=2.4.0
tensorboard>=2.7.0
pycocotools>=2.0
onnx>=1.10.0  # ONNX export
onnx-simplifier>=0.3.6 # ONNX simplifier
thop  # FLOPs computation
#pytorch_quantization>=2.1.1
